@@ -1,30 +1,59 @@
-require 'bundler/capistrano'
-load 'deploy/assets'
-set :application, "EntryApplication"
-set :repository,  "https://github.com/jeremiaheb/EntryApplication.git"
+# config valid only for Capistrano 3.1
+lock '3.8.1'
 
-set :scm, :git
-set :deploy_via, :remote_cache
+set :application, 'EntryApplication'
+set :repo_url, 'git@github.com:jeremiaheb/EntryApplication.git'
+set :branch , ENV['BRANCH'] if ENV['BRANCH']
+# Default branch is :master
+# ask :branch, proc { `git rev-parse --abbrev-ref HEAD`.chomp }
 
-set :use_sudo, false
-set :user, 'jeremiaheb'
-set :branch, :master
-set :rails_env, :production
-set :deploy_to, "/var/www/apps/#{application}"
-role :web, "199.242.232.49"                          # Your HTTP server, Apache/etc
-role :app, "199.242.232.49"                          # This may be the same as your `Web` server
-role :db,  "199.242.232.49", :primary => true # This is where Rails migrations will run
+# Default deploy_to directory is /var/www/my_app
+set :deploy_to, '/var/www/apps/EntryApplication'
 
+# Default value for :scm is :git
+#set :scm, :git
 
-default_run_options[:pty] = true
-#ssh_options[:keys] = [File.join(ENV["HOME"], ".ssh", "id_rsa")]
-ssh_options[:forward_agent] = true
+# Default value for :format is :pretty
+# set :format, :pretty
 
-# If you are using Passenger mod_rails uncomment this:
+# Default value for :log_level is :debug
+# set :log_level, :debug
+
+# Default value for :pty is false
+# set :pty, true
+
+# Default value for :linked_files is []
+set :linked_files, %w{config/database.yml}
+
+# Default value for linked_dirs is []
+set :linked_dirs, %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system public/data}
+
+# Default value for default_env is {}
+# set :default_env, { path: "/opt/ruby/bin:$PATH" }
+
+# Default value for keep_releases is 5
+# set :keep_releases, 5
+
 namespace :deploy do
-  task :start do ; end
-  task :stop do ; end
-  task :restart, :roles => :app, :except => { :no_release => true } do
-    run "#{try_sudo} touch #{File.join(current_path,'tmp','restart.txt')}"
+
+  desc 'Restart application'
+  task :restart do
+    on roles(:app), in: :sequence, wait: 5 do
+      # Your restart mechanism here, for example:
+      execute :touch, release_path.join('tmp/restart.txt')
+    end
   end
+
+  after :publishing, :restart
+  after :finishing,  :restart
+
+  after :restart, :clear_cache do
+    on roles(:web), in: :groups, limit: 3, wait: 10 do
+      # Here we can do anything such as:
+      # within release_path do
+      #   execute :rake, 'cache:clear'
+      # end
+    end
+  end
+
 end
