@@ -11,18 +11,18 @@
 // GO AFTER THE REQUIRES BELOW.
 //
 //= require jquery
+//= require jquery-ui
 //= require jquery_ujs
-//= require twitter/bootstrap
-//= require jquery.ui.datepicker
 //= require dataTables/jquery.dataTables
-//= require dataTables/jquery.dataTables.bootstrap
-//= require jquery_nested_form
-//= require rails.validations
-//= require rails.validations.nested_form
+//= require dataTables/bootstrap/3/jquery.dataTables.bootstrap
+//= require bootstrap-datepicker
 //= require select2
+//= require jquery.validate
+//= require twitter/bootstrap
+//= require jquery-ui/datepicker
+//= require jquery_nested_form
 //= require_tree ../../../vendor/assets/javascripts
 //= require_tree .
-//= require blueimp-gallery-all
 //
 
 // Set up our EA namespace for our functions
@@ -55,7 +55,7 @@ $(function(){
     $('.display').dataTable( {
         //"sDom": "<'row'<'span7'lf>r>t<'row'<'span7'ip>>",
         "sDom": '<"top"ifp<"clear">>rt<"bottom"<"clear">>',
-        "sPaginationType": "bootstrap"
+        "pagingType": "full"
     } );
 
     $(document).ready(function(){
